in_gpu_metrics: Unify include/exclude pattern accross AMD/NVDA GPUs
